+++ src/fonts/open-sans-v40-latin-600.woff2
