+++ src/js/app.js
@@ -1,0 +1,15 @@
+/**
+ * The main js file for this theme.
+ * You can add both internal and external dependencies.
+ * Webpack will take care of the rest
+ * Add your scripts here.
+ */
+
+/**
+ * Internal Dependencies
+ */
+import { initLazyLoadFallback } from './lazy-load-fallback';
+
+// Initialize lazy load fallback
+initLazyLoadFallback();
+
